+++ _modules/vectra/lib/MiniMaxEmbeddings.spec.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=MiniMaxEmbeddings.spec.d.ts.map